Moved all xarray dependencies into _xr_interop
Can now import GeoBox without xarray.
